feat: tx_origin in events, refactor
