+++ async-scheduler/src/events_extractor/models.rs
@@ -1,22 +1,6 @@
 use bigdecimal::BigDecimal;
 use chrono::NaiveDateTime;
 
-#[derive(Debug)]
-pub struct Blockscan {
-    pub base_url: String,
-    pub block_from: String,
-    pub block_to: String,
-    pub address: String,
-    pub topic0: String,
-    pub is_topic1: bool,
-    pub topic1: String,
-    pub is_topic2: bool,
-    pub topic2: String,
-    pub is_topic3: bool,
-    pub topic3: String,
-    pub apikey: String,
-}
-
 pub struct Event {
     // pub address: String,
     pub topics: Vec<String>,
@@ -32,6 +16,7 @@ pub struct Event {
 
 #[derive(Debug)]
 pub struct EventERC20Transfer {
+    pub tx_origin: String,
     pub sender: String,
     pub receiver: String,
     pub amount: BigDecimal,
@@ -43,6 +28,7 @@ pub struct EventERC20Transfer {
 
 #[derive(Debug)]
 pub struct EventERC20Approval {
+    pub tx_origin: String,
     pub owner: String,
     pub spender: String,
     pub amount: BigDecimal,
@@ -54,6 +40,7 @@ pub struct EventERC20Approval {
 
 #[derive(Debug)]
 pub struct EventAnyV4Swapin {
+    pub tx_origin: String,
     pub account: String,
     pub amount: BigDecimal,
     pub transfer_tx_hash: String,
@@ -65,6 +52,7 @@ pub struct EventAnyV4Swapin {
 
 #[derive(Debug)]
 pub struct EventAnyV4Swapout {
+    pub tx_origin: String,
     pub account: String,
     pub bindaddr: String,
     pub amount: BigDecimal,
@@ -76,6 +64,7 @@ pub struct EventAnyV4Swapout {
 
 #[derive(Debug)]
 pub struct EventUniV2PairCreated {
+    pub tx_origin: String,
     pub address: String,
     pub token0: String,
     pub token1: String,
@@ -90,6 +79,7 @@ pub struct EventUniV2PairCreated {
 
 #[derive(Debug)]
 pub struct EventUniV2Swap {
+    pub tx_origin: String,
     pub sender: String,
     pub receiver: String,
     pub amount0_in: BigDecimal,
@@ -104,6 +94,7 @@ pub struct EventUniV2Swap {
 
 #[derive(Debug)]
 pub struct EventUniV2Mint {
+    pub tx_origin: String,
     pub sender: String,
     pub amount0: BigDecimal,
     pub amount1: BigDecimal,
@@ -115,6 +106,7 @@ pub struct EventUniV2Mint {
 
 #[derive(Debug)]
 pub struct EventUniV2Burn {
+    pub tx_origin: String,
     pub sender: String,
     pub receiver: String,
     pub amount0: BigDecimal,
@@ -123,4 +115,57 @@ pub struct EventUniV2Burn {
     pub block_number: i64,
     pub tx_hash: String,
     pub log_index: i64,
+}
+
+#[derive(Debug)]
+pub struct UniV2Sell {
+    pub swap_id: i64,
+    pub pair_id: i64,
+    pub pair_title: String,
+    pub tx_origin: String,
+    pub amount_gton_in: BigDecimal,
+    pub amount_token_out: BigDecimal,
+    pub stamp: NaiveDateTime,
+    pub tx_hash: String,
+    pub log_index: i64
+}
+
+#[derive(Debug)]
+pub struct UniV2Buy {
+    pub swap_id: i64,
+    pub pair_id: i64,
+    pub pair_title: String,
+    pub tx_origin: String,
+    pub amount_token_in: BigDecimal,
+    pub amount_gton_out: BigDecimal,
+    pub stamp: NaiveDateTime,
+    pub tx_hash: String,
+    pub log_index: i64
+}
+
+#[derive(Debug)]
+pub struct UniV2LPAdd {
+    pub mint_id: i64,
+    pub pair_id: i64,
+    pub pair_title: String,
+    pub tx_origin: String,
+    pub amount_gton_in: BigDecimal,
+    pub amount_token_out: BigDecimal,
+    pub amount_lp_out: BigDecimal,
+    pub stamp: NaiveDateTime,
+    pub tx_hash: String,
+    pub log_index: i64
+}
+
+#[derive(Debug)]
+pub struct UniV2LPRemove {
+    pub burn_id: i64,
+    pub pair_id: i64,
+    pub pair_title: String,
+    pub tx_origin: String,
+    pub amount_gton_in: BigDecimal,
+    pub amount_token_out: BigDecimal,
+    pub stamp: NaiveDateTime,
+    pub tx_hash: String,
+    pub log_index: i64
 }
