+++ async-scheduler/src/events_extractor/pollers.rs
@@ -1,150 +1,135 @@
+use anyhow::{Context, Result};
+
+use web3::transports::Http;
+use web3::contract::{Contract, Options};
+use web3::{
+    Web3,
+    types::{
+        Block,
+        Address,
+        Log,
+        FilterBuilder,
+        BlockNumber,
+        U256,
+        H256,
+        U64,
+        TransactionId
+    }
+};
+use web3::ethabi::{
+    Topic,
+    TopicFilter,
+};
+use std::convert::TryFrom;
+use std::str::FromStr;
+use std::ops::Index;
+use hex::ToHex;
+
 use crate::DbPool;
 use diesel::prelude::*;
-use diesel::sql_types::{BigInt, Numeric, Text, Timestamp};
+use diesel::sql_types::{BigInt, Bool, Numeric, Text, Timestamp};
 use diesel::result::Error::DatabaseError;
 use diesel::result::DatabaseErrorKind::UniqueViolation;
 
+use bigdecimal::{BigDecimal, ToPrimitive};
+
+use chrono::NaiveDateTime;
+
 use super::constants::C;
-use super::fetchers::{fetch_events_erc20_approval, fetch_events_erc20_transfer, fetch_events_anyv4_transfer, fetch_events_anyv4_swapin, fetch_events_anyv4_swapout, fetch_events_univ2_pair_created, fetch_events_univ2_swap, fetch_events_univ2_mint, fetch_events_univ2_burn};
-use super::parsers::{parse_events_erc20_approval, parse_events_erc20_transfer, parse_events_anyv4_swapin, parse_events_anyv4_swapout, parse_events_univ2_pair_created, parse_events_univ2_swap, parse_events_univ2_mint, parse_events_univ2_burn};
+use super::models::*;
 
 #[derive(QueryableByName, PartialEq, Debug)]
-struct BlockNumber {
+struct LastBlock {
     #[sql_type = "BigInt"]
     block_number: i64,
 }
 
-#[derive(QueryableByName, PartialEq, Debug)]
-struct Pair {
-    #[sql_type = "BigInt"]
-    id: i64,
-    #[sql_type = "Text"]
-    address: String,
+async fn fetch_stamp(web3: &web3::Web3<Http>, block_number: &Option<U64>) -> NaiveDateTime {
+   let block: Block<H256> = web3.eth().block(BlockNumber::Number(block_number.unwrap()).into()).await.unwrap().unwrap();
+   let stamp_str = block.timestamp.to_string();
+   let stamp_big = BigDecimal::from_str(&stamp_str).unwrap();
+   let stamp_i64 = stamp_big.to_i64().unwrap();
+   NaiveDateTime::from_timestamp(stamp_i64,0)
 }
 
-pub async fn poll_events(p: &DbPool, ftmscan_api_key: &str, ethscan_api_key: &str) {
+fn parse_block_number(block_number: &Option<U64>) -> i64 {
+   let block_number_str = block_number.unwrap().to_string();
+   let block_number_big = BigDecimal::from_str(&block_number_str).unwrap();
+   block_number_big.to_i64().unwrap()
+}
 
-    poll_events_erc20_approval(
-        p,
-        "events_erc20_approval_ftm",
-        "https://api.ftmscan.com",
-        ftmscan_api_key,
-        C.ftm_gton
-    ).await;
-
-    poll_events_erc20_transfer(
-        p,
-        "events_erc20_transfer_ftm",
-        "https://api.ftmscan.com",
-        ftmscan_api_key,
-        C.ftm_gton
-    ).await;
-
-    poll_events_anyv4_transfer(
-        p,
-        ethscan_api_key
-    ).await;
-
-    poll_events_anyv4_swapin(
-        p,
-        "events_anyv4_swapin_ftm",
-        "https://api.ftmscan.com",
-        ftmscan_api_key,
-        C.ftm_gton
-    ).await;
-
-    poll_events_anyv4_swapout(
-        p,
-        "events_anyv4_swapout_ftm",
-        "https://api.ftmscan.com",
-        ftmscan_api_key,
-        C.ftm_gton
-    ).await;
-
-    poll_events_univ2_pair_created(
-        p,
-        "events_univ2_pair_created_spirit",
-        "https://api.ftmscan.com",
-        ftmscan_api_key,
-        C.ftm_gton,
-        C.ftm_spirit_factory
-    ).await;
-
-    let pairs = diesel::sql_query(
-        format!("SELECT id, address FROM {} ORDER BY block_number DESC;", "events_univ2_pair_created_spirit"),
-    )
-        .get_results::<Pair>(&p.get().unwrap()).unwrap();
-
-    for pair in pairs {
-        poll_events_univ2_swap(
-            p,
-            "events_univ2_swap_spirit",
-            "https://api.ftmscan.com",
-            ftmscan_api_key,
-            pair.id,
-            &pair.address,
-        ).await;
-        poll_events_univ2_mint(
-            p,
-            "events_univ2_mint_spirit",
-            "https://api.ftmscan.com",
-            ftmscan_api_key,
-            pair.id,
-            &pair.address,
-        ).await;
-        poll_events_univ2_burn(
-            p,
-            "events_univ2_burn_spirit",
-            "https://api.ftmscan.com",
-            ftmscan_api_key,
-            pair.id,
-            &pair.address,
-        ).await;
-    }
+fn hex_to_string<T: ToHex>(h: T) -> String {
+    "0x".to_owned() + &h.encode_hex::<String>()
 }
 
 pub async fn poll_events_erc20_approval(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
-    gton: &str
+    web3: &web3::Web3<Http>,
+    token: &str,
 ) {
 
     println!("polling events erc20 approval");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
+    let last_block: BlockNumber = match diesel::sql_query(
         format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_erc20_approval = fetch_events_erc20_approval(
-        api_url,
-        gton,
-        &last_block,
-        api_key,
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_erc20_approve.parse().unwrap());
 
-    let events_erc20_approval = parse_events_erc20_approval(values_erc20_approval)
-    .await
-    .unwrap();
-    for event in events_erc20_approval {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![token.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+
+    let mut events: Vec<EventERC20Approval> = vec![];
+    for e in result.into_iter() {
+        let owner: String = hex_to_string(Address::from(e.topics[1]));
+        let spender: String = hex_to_string(Address::from(e.topics[2]));
+        let amount: BigDecimal = BigDecimal::from_str(&U256::from_big_endian(&e.data.0).to_string()).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventERC20Approval {
+            tx_origin,
+            owner,
+            spender,
+            amount,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
+                               tx_origin,\
                                owner,\
                                spender,\
                                amount,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
             table_name),
         )
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.owner)
         .bind::<Text, _>(&event.spender)
         .bind::<Numeric, _>(&event.amount)
@@ -152,7 +137,7 @@ pub async fn poll_events_erc20_approval(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -163,47 +148,72 @@ pub async fn poll_events_erc20_approval(
 }
 
 pub async fn poll_events_erc20_transfer(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
-    gton: &str
+    web3: &web3::Web3<Http>,
+    token: &str,
 ) {
 
     println!("polling events erc20 transfer");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
+    let last_block: BlockNumber = match diesel::sql_query(
         format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_erc20_transfer = fetch_events_erc20_transfer(
-        api_url,
-        gton,
-        &last_block,
-        api_key,
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_erc20_transfer.parse().unwrap());
 
-    let events_erc20_transfer = parse_events_erc20_transfer(values_erc20_transfer)
-    .await
-    .unwrap();
-    for event in events_erc20_transfer {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![token.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventERC20Transfer> = vec![];
+    for e in result.into_iter() {
+        let sender: String = hex_to_string(Address::from(e.topics[1]));
+        let receiver: String = hex_to_string(Address::from(e.topics[2]));
+        let amount: BigDecimal = BigDecimal::from_str(&U256::from_big_endian(&e.data.0).to_string()).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventERC20Transfer {
+            tx_origin,
+            sender,
+            receiver,
+            amount,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
+                               tx_origin,\
                                sender,\
                                receiver,\
                                amount,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
             table_name),
         )
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.sender)
         .bind::<Text, _>(&event.receiver)
         .bind::<Numeric, _>(&event.amount)
@@ -211,7 +221,7 @@ pub async fn poll_events_erc20_transfer(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -222,44 +232,73 @@ pub async fn poll_events_erc20_transfer(
 }
 
 pub async fn poll_events_anyv4_transfer(
-    p: &DbPool,
-    api_key: &str,
+    pool: &DbPool,
+    table_name: &str,
+    web3: &web3::Web3<Http>,
+    token: &str,
 ) {
+
     println!("polling events anyv4 transfer");
-
-    let table_name = "events_anyv4_transfer";
-
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
+    let last_block: BlockNumber = match diesel::sql_query(
         format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_erc20_transfer = fetch_events_anyv4_transfer(
-        &last_block,
-        api_key,
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_erc20_transfer.parse().unwrap());
+    topics.topic2 = Topic::This(C.eth_anyv4_vault.parse().unwrap());
 
-    let events_erc20_transfer = parse_events_erc20_transfer(values_erc20_transfer)
-    .await
-    .unwrap();
-    for event in events_erc20_transfer {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![token.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventERC20Transfer> = vec![];
+    for e in result.into_iter() {
+        let sender: String = hex_to_string(Address::from(e.topics[1]));
+        let receiver: String = hex_to_string(Address::from(e.topics[2]));
+        let amount: BigDecimal = BigDecimal::from_str(&U256::from_big_endian(&e.data.0).to_string()).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventERC20Transfer {
+            tx_origin,
+            sender,
+            receiver,
+            amount,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
+                               tx_origin,\
                                sender,\
                                receiver,\
                                amount,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
             table_name),
         )
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.sender)
         .bind::<Text, _>(&event.receiver)
         .bind::<Numeric, _>(&event.amount)
@@ -267,7 +306,7 @@ pub async fn poll_events_anyv4_transfer(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -278,47 +317,72 @@ pub async fn poll_events_anyv4_transfer(
 }
 
 pub async fn poll_events_anyv4_swapin(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
-    gton: &str
+    web3: &web3::Web3<Http>,
+    token: &str,
 ) {
 
     println!("polling events anyv4 swapin");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
+    let last_block: BlockNumber = match diesel::sql_query(
         format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_anyv4_swapin = fetch_events_anyv4_swapin(
-        api_url,
-        gton,
-        &last_block,
-        api_key,
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_anyv4_swapin.parse().unwrap());
 
-    let events_anyv4_swapin = parse_events_anyv4_swapin(values_anyv4_swapin)
-    .await
-    .unwrap();
-    for event in events_anyv4_swapin {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![token.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventAnyV4Swapin> = vec![];
+    for e in result.into_iter() {
+        let transfer_tx_hash: String = hex_to_string(e.topics[1]);
+        let account: String = hex_to_string(Address::from(e.topics[2]));
+        let amount: BigDecimal = BigDecimal::from_str(&U256::from_big_endian(&e.data.0).to_string()).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventAnyV4Swapin {
+            tx_origin,
+            transfer_tx_hash,
+            account,
+            amount,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
+                               tx_origin,\
                                account,\
                                amount,\
                                transfer_tx_hash,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
             table_name),
         )
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.account)
         .bind::<Numeric, _>(&event.amount)
         .bind::<Text, _>(&event.transfer_tx_hash)
@@ -326,7 +390,7 @@ pub async fn poll_events_anyv4_swapin(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -337,47 +401,72 @@ pub async fn poll_events_anyv4_swapin(
 }
 
 pub async fn poll_events_anyv4_swapout(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
-    gton: &str
+    web3: &web3::Web3<Http>,
+    token: &str,
 ) {
 
     println!("polling events anyv4 swapout");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
+    let last_block: BlockNumber = match diesel::sql_query(
         format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_anyv4_swapout = fetch_events_anyv4_swapout(
-        api_url,
-        gton,
-        &last_block,
-        api_key,
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_anyv4_swapout.parse().unwrap());
 
-    let events_anyv4_swapout = parse_events_anyv4_swapout(values_anyv4_swapout)
-    .await
-    .unwrap();
-    for event in events_anyv4_swapout {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![token.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventAnyV4Swapout> = vec![];
+    for e in result.into_iter() {
+        let account: String = hex_to_string(Address::from(e.topics[1]));
+        let bindaddr: String = hex_to_string(Address::from(e.topics[2]));
+        let amount: BigDecimal = BigDecimal::from_str(&U256::from_big_endian(&e.data.0).to_string()).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventAnyV4Swapout {
+            tx_origin,
+            account,
+            bindaddr,
+            amount,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
+                               tx_origin,\
                                account,\
                                bindaddr,\
                                amount,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
             table_name),
         )
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.account)
         .bind::<Text, _>(&event.bindaddr)
         .bind::<Numeric, _>(&event.amount)
@@ -385,7 +474,7 @@ pub async fn poll_events_anyv4_swapout(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -395,60 +484,134 @@ pub async fn poll_events_anyv4_swapout(
     }
 }
 
+async fn get_token_name (web3: &Web3<Http>, token: Address) -> String {
+    let contract = Contract::from_json(
+        web3.eth(),
+        token,
+        include_bytes!("abi/erc20.json"),
+    )
+    .expect("create erc20 contract");
+
+    let res: String = contract
+        .query("name", (), None, Options::default(), None)
+        .await
+        .expect("get token name");
+    res.to_string()
+}
+
 pub async fn poll_events_univ2_pair_created(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
-    gton: &str,
+    web3: &web3::Web3<Http>,
+    token: &str,
     factory: &str
 ) {
 
     println!("polling events univ2 pair created");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
+    let last_block: BlockNumber = match diesel::sql_query(
         format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_univ2_pair_created = fetch_events_univ2_pair_created(
-        api_url,
-        gton,
-        factory,
-        &last_block,
-        api_key
-    )
-    .await.unwrap();
+    let token_addr: Address = token.parse().unwrap();
+    let token_h256 = H256::from(token_addr);
 
-    let events_univ2_pair_created = parse_events_univ2_pair_created(values_univ2_pair_created, gton)
-    .await
-    .unwrap();
-    for event in events_univ2_pair_created {
+    // check pairs where gton is token0
+    let mut topics1 = TopicFilter::default();
+    topics1.topic0 = Topic::This(C.topic0_univ2_pair_created.parse().unwrap());
+    topics1.topic1 = Topic::This(token_h256);
+
+    let filter1 = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![factory.parse().unwrap()])
+                .topic_filter(topics1)
+                .build();
+    let result1: Vec<web3::types::Log> = web3.eth().logs(filter1).await.unwrap();
+
+    // check pairs where gton is token1
+    let mut topics2 = TopicFilter::default();
+    topics2.topic0 = Topic::This(C.topic0_univ2_pair_created.parse().unwrap());
+    topics2.topic2 = Topic::This(token_h256);
+
+    let filter2 = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![factory.parse().unwrap()])
+                .topic_filter(topics2)
+                .build();
+    let result2: Vec<web3::types::Log> = web3.eth().logs(filter2).await.unwrap();
+
+    let result = [result1, result2].concat();
+
+    let mut events: Vec<EventUniV2PairCreated> = vec![];
+    for e in result.into_iter() {
+        let token0: String = hex_to_string(Address::from(e.topics[1]));
+        let token1: String = hex_to_string(Address::from(e.topics[2]));
+        let mut address_bytes: [u8; 20] = [0;20];
+        address_bytes.copy_from_slice(&e.data.0[12..32]);
+        let address = hex_to_string(Address::from(&address_bytes));
+        let gtonToken0: bool = token0 == token;
+        let title0 = get_token_name(&web3, Address::from(e.topics[1])).await;
+        let title1 = get_token_name(&web3, Address::from(e.topics[2])).await;
+        let title = title0 + "-" + &title1;
+        let index_str = U256::from(&e.data.0[32..64]).to_string();
+        let index = BigDecimal::from_str(&index_str).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventUniV2PairCreated {
+            tx_origin,
+            address,
+            token0,
+            token1,
+            index,
+            gtonToken0,
+            title,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
+                               tx_origin,\
                                address,\
                                token0,\
                                token1,\
+                               gtonToken0,\
+                               title,\
                                index,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11);",
             table_name),
         )
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.address)
         .bind::<Text, _>(&event.token0)
         .bind::<Text, _>(&event.token1)
+        .bind::<Bool, _>(&event.gtonToken0)
+        .bind::<Text, _>(&event.title)
         .bind::<Numeric, _>(&event.index)
         .bind::<Timestamp, _>(&event.stamp)
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -458,41 +621,167 @@ pub async fn poll_events_univ2_pair_created(
     }
 }
 
-pub async fn poll_events_univ2_swap(
-    p: &DbPool,
+pub async fn poll_events_univ2_transfer(
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
+    web3: &web3::Web3<Http>,
+    pair_id: i64,
+    pair_address: &str,
+) {
+
+    println!("polling events lp transfer");
+    // get latest block from db
+    let last_block: BlockNumber = match diesel::sql_query(
+        format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
+    )
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
+        };
+    println!("starting from block {:#?}", last_block);
+
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_erc20_transfer.parse().unwrap());
+
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![pair_address.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventERC20Transfer> = vec![];
+    for e in result.into_iter() {
+        let sender: String = hex_to_string(Address::from(e.topics[1]));
+        let receiver: String = hex_to_string(Address::from(e.topics[2]));
+        let amount: BigDecimal = BigDecimal::from_str(&U256::from_big_endian(&e.data.0).to_string()).unwrap();
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventERC20Transfer {
+            tx_origin,
+            sender,
+            receiver,
+            amount,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
+        let result = diesel::sql_query(
+            format!("insert into {}(\
+                               pair_id,\
+                               tx_origin,\
+                               sender,\
+                               receiver,\
+                               amount,\
+                               stamp,\
+                               block_number,\
+                               tx_hash,
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9);",
+            table_name),
+        )
+        .bind::<BigInt, _>(&pair_id)
+        .bind::<Text, _>(&event.tx_origin)
+        .bind::<Text, _>(&event.sender)
+        .bind::<Text, _>(&event.receiver)
+        .bind::<Numeric, _>(&event.amount)
+        .bind::<Timestamp, _>(&event.stamp)
+        .bind::<BigInt, _>(&event.block_number)
+        .bind::<Text, _>(&event.tx_hash)
+        .bind::<BigInt, _>(&event.log_index)
+        .execute(&pool.get().unwrap());
+        match result {
+            // ignore if already processed, panic otherwise
+            Ok(_) => continue,
+            Err(DatabaseError(UniqueViolation, _)) => continue,
+            Err(e) => panic!("write to db: {:#?}, err {}", &event, e),
+        };
+    }
+}
+pub async fn poll_events_univ2_swap(
+    pool: &DbPool,
+    table_name: &str,
+    web3: &web3::Web3<Http>,
     pair_id: i64,
     pair_address: &str
 ) {
 
     println!("polling events univ2 swap");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
-        format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
+    let last_block: BlockNumber = match diesel::sql_query(
+        format!("SELECT block_number FROM {} \
+                 WHERE pair = {} \
+                 ORDER BY block_number DESC;",
+                table_name,
+                pair_id),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_univ2_swap = fetch_events_univ2_swap(
-        api_url,
-        pair_address,
-        &last_block,
-        api_key
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_univ2_swap.parse().unwrap());
 
-    let events_univ2_swap = parse_events_univ2_swap(values_univ2_swap)
-    .await
-    .unwrap();
-    for event in events_univ2_swap {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![pair_address.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventUniV2Swap> = vec![];
+    for e in result.into_iter() {
+        let sender: String = hex_to_string(Address::from(e.topics[1]));
+        let receiver: String = hex_to_string(Address::from(e.topics[2]));
+
+        let amount0_in_str = U256::from(&e.data.0[..32]).to_string();
+        let amount0_in = BigDecimal::from_str(&amount0_in_str).unwrap();
+        let amount1_in_str = U256::from(&e.data.0[32..64]).to_string();
+        let amount1_in = BigDecimal::from_str(&amount1_in_str).unwrap();
+        let amount0_out_str = U256::from(&e.data.0[64..96]).to_string();
+        let amount0_out = BigDecimal::from_str(&amount0_out_str).unwrap();
+        let amount1_out_str = U256::from(&e.data.0[96..128]).to_string();
+        let amount1_out = BigDecimal::from_str(&amount1_out_str).unwrap();
+
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventUniV2Swap {
+            tx_origin,
+            sender,
+            receiver,
+            amount0_in,
+            amount1_in,
+            amount0_out,
+            amount1_out,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
-                               pair,
+                               pair_id,\
+                               tx_origin,\
                                sender,\
                                receiver,\
                                amount0_in,\
@@ -502,10 +791,11 @@ pub async fn poll_events_univ2_swap(
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12);",
             table_name),
         )
         .bind::<BigInt, _>(pair_id)
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.sender)
         .bind::<Text, _>(&event.receiver)
         .bind::<Numeric, _>(&event.amount0_in)
@@ -516,7 +806,7 @@ pub async fn poll_events_univ2_swap(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -527,50 +817,81 @@ pub async fn poll_events_univ2_swap(
 }
 
 pub async fn poll_events_univ2_mint(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
+    web3: &web3::Web3<Http>,
     pair_id: i64,
     pair_address: &str
 ) {
 
     println!("polling events univ2 mint");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
-        format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
+    let last_block: BlockNumber = match diesel::sql_query(
+        format!("SELECT block_number FROM {} \
+                 WHERE pair = {} \
+                 ORDER BY block_number DESC;", table_name, pair_id),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_univ2_mint = fetch_events_univ2_mint(
-        api_url,
-        pair_address,
-        &last_block,
-        api_key
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_univ2_mint.parse().unwrap());
 
-    let events_univ2_mint = parse_events_univ2_mint(values_univ2_mint)
-    .await
-    .unwrap();
-    for event in events_univ2_mint {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![pair_address.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventUniV2Mint> = vec![];
+    for e in result.into_iter() {
+        let sender: String = hex_to_string(Address::from(e.topics[1]));
+
+        let amount0_str = U256::from(&e.data.0[..32]).to_string();
+        let amount0 = BigDecimal::from_str(&amount0_str).unwrap();
+        let amount1_str = U256::from(&e.data.0[32..64]).to_string();
+        let amount1 = BigDecimal::from_str(&amount1_str).unwrap();
+
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventUniV2Mint {
+            tx_origin,
+            sender,
+            amount0,
+            amount1,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
-                               pair,
+                               pair_id,\
+                               tx_origin,\
                                sender,\
                                amount0,\
                                amount1,\
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9);",
             table_name),
         )
         .bind::<BigInt, _>(pair_id)
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.sender)
         .bind::<Numeric, _>(&event.amount0)
         .bind::<Numeric, _>(&event.amount1)
@@ -578,7 +899,7 @@ pub async fn poll_events_univ2_mint(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
@@ -589,40 +910,72 @@ pub async fn poll_events_univ2_mint(
 }
 
 pub async fn poll_events_univ2_burn(
-    p: &DbPool,
+    pool: &DbPool,
     table_name: &str,
-    api_url: &str,
-    api_key: &str,
+    web3: &web3::Web3<Http>,
     pair_id: i64,
     pair_address: &str
 ) {
 
     println!("polling events univ2 burn");
     // get latest block from db
-    let last_block: String = match diesel::sql_query(
-        format!("SELECT block_number FROM {} ORDER BY block_number DESC;", table_name),
+    let last_block: BlockNumber = match diesel::sql_query(
+        format!("SELECT block_number FROM {} \
+                 WHERE pair = {} \
+                 ORDER BY block_number DESC;", table_name, pair_id),
     )
-        .get_result::<BlockNumber>(&p.get().unwrap()) {
-            Err(_) => "earliest".to_string(),
-            Ok(e) => e.block_number.to_string()
+        .get_result::<LastBlock>(&pool.get().unwrap()) {
+            Err(_) => BlockNumber::Earliest,
+            Ok(e) => BlockNumber::Number(e.block_number.into())
         };
-    println!("starting from block {}", last_block);
+    println!("starting from block {:#?}", last_block);
 
-    let values_univ2_burn = fetch_events_univ2_burn(
-        api_url,
-        pair_address,
-        &last_block,
-        api_key
-    )
-    .await.unwrap();
+    let mut topics = TopicFilter::default();
+    topics.topic0 = Topic::This(C.topic0_univ2_burn.parse().unwrap());
 
-    let events_univ2_burn = parse_events_univ2_burn(values_univ2_burn)
-    .await
-    .unwrap();
-    for event in events_univ2_burn {
+    let filter = FilterBuilder::default()
+                .from_block(last_block)
+                .to_block(BlockNumber::Latest)
+                .address(vec![pair_address.parse().unwrap()])
+                .topic_filter(topics)
+                .build();
+    let result: Vec<web3::types::Log> = web3.eth().logs(filter).await.unwrap();
+
+    let mut events: Vec<EventUniV2Burn> = vec![];
+    for e in result.into_iter() {
+        let sender: String = hex_to_string(Address::from(e.topics[1]));
+        let receiver: String = hex_to_string(Address::from(e.topics[2]));
+
+        let amount0_str = U256::from(&e.data.0[..32]).to_string();
+        let amount0 = BigDecimal::from_str(&amount0_str).unwrap();
+        let amount1_str = U256::from(&e.data.0[32..64]).to_string();
+        let amount1 = BigDecimal::from_str(&amount1_str).unwrap();
+
+        let block_number = parse_block_number(&e.block_number);
+        let stamp = fetch_stamp(&web3, &e.block_number).await;
+        let tx_hash = hex_to_string(e.transaction_hash.unwrap());
+        let log_index = i64::try_from(e.log_index.unwrap().as_u64()).unwrap();
+        // get transaction origin
+        let tx = &web3.eth().transaction(TransactionId::Hash(tx_hash.parse().unwrap())).await.unwrap().unwrap();
+        let tx_origin = hex_to_string(tx.from);
+        events.push(EventUniV2Burn {
+            tx_origin,
+            sender,
+            receiver,
+            amount0,
+            amount1,
+            stamp,
+            block_number,
+            tx_hash,
+            log_index
+        })
+    }
+
+    for event in events {
         let result = diesel::sql_query(
             format!("insert into {}(\
-                               pair,
+                               pair_id,\
+                               tx_origin,\
                                sender,\
                                receiver,\
                                amount0,\
@@ -630,10 +983,11 @@ pub async fn poll_events_univ2_burn(
                                stamp,\
                                block_number,\
                                tx_hash,
-                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9);",
+                               log_index) VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10);",
             table_name),
         )
         .bind::<BigInt, _>(pair_id)
+        .bind::<Text, _>(&event.tx_origin)
         .bind::<Text, _>(&event.sender)
         .bind::<Text, _>(&event.receiver)
         .bind::<Numeric, _>(&event.amount0)
@@ -642,7 +996,7 @@ pub async fn poll_events_univ2_burn(
         .bind::<BigInt, _>(&event.block_number)
         .bind::<Text, _>(&event.tx_hash)
         .bind::<BigInt, _>(&event.log_index)
-        .execute(&p.get().unwrap());
+        .execute(&pool.get().unwrap());
         match result {
             // ignore if already processed, panic otherwise
             Ok(_) => continue,
