+++ async-scheduler/src/main.rs
@@ -4,7 +4,6 @@ use async_scheduler::price_coingecko::CoingeckoPrice;
 use async_scheduler::keeper_extractor::KeeperExtractor;
 use async_scheduler::forum_extractor::ForumExtractor;
 use async_scheduler::events_extractor::EventsExtractor;
-use async_scheduler::events_reports::ReportsExtractor;
 
 use diesel::r2d2::{ConnectionManager,Pool};
 use diesel::PgConnection;
@@ -24,12 +23,6 @@ async fn main() {
     tokio::task::spawn(async move {
         EventsExtractor::new(p.clone()).run().await;
     }).await.unwrap();
-
-    let p = pool.clone();
-    tokio::task::spawn(async move {
-        ReportsExtractor::new(p.clone()).run().await;
-    }).await.unwrap();
-
 
     // let p = pool.clone();
     // tokio::task::spawn(async move {
